convert vdom on performance webapp
