+++ tugas/webapp/www/performance.html
@@ -7,13 +7,15 @@
     <title>Task Manager</title>
   </head>
   <body>
-    <ul>
+    <!-- dom -->
+    <!-- <ul>
       <li><a href="worker.html">pekerja</a></li>
       <li><a href="tasks.html">pekerjaan</a></li>
       <li><a href="performance.html">kinerja</a></li>
     </ul>
-    <hr />
-    <p id="error-text" class="error"></p>
+    <hr /> -->
+
+    <!-- <p id="error-text" class="error"></p>
     <p id="loading-text" class="primary">memuat...</p>
     <button id="refresh">refresh</button>
     <ul>
@@ -21,7 +23,10 @@
       <li>jumlah tugas: <span id="tasks"></span></li>
       <li>yang selesai: <span id="task-done"></span></li>
       <li>yang diabtalkan: <span id="task-canceled"></span></li>
-    </ul>
+    </ul> -->
+
+    <!-- vdom -->
+    <div id="performance"></div>
   </body>
   <script type="text/javascript" src="performance.js"></script>
 </html>
