+++ tugas/webpack.config.js
@@ -24,15 +24,6 @@ module.exports = {
         test: /\.css$/i,
         use: ['style-loader', 'css-loader'],
       },
-      {
-        test: /\.yaml$/,
-        use: [{ loader: 'json-loader' }, { loader: 'yaml-loader' }],
-      },
-      {
-        test: /\.ts$/,
-        loader: 'ts-loader',
-        exclude: /node_modules/,
-      },
     ],
   },
 };
